requirements.txt
requests~=2.32.3
Flask~=3.0.3
GitPython~=3.1.43
gpt4all~=2.8.2